requirements.txt
alembic~=1.11.1
anyio==3.7.0
click==8.1.3
Deprecated==1.2.14
python-dotenv==1.0.0
exceptiongroup==1.1.1
fastapi~=0.95.1
geoalchemy2~=0.14.1
greenlet~=2.0.2
gunicorn~=21.2.0
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
limits==3.5.0
numpy~=1.25.1
orjson~=3.9.2
packaging==23.1
psycopg~=3.1.10
psycopg-binary~=3.1.10
pydantic~=1.10.8
shapely~=2.0.1
slowapi~=0.1.8
sniffio==1.3.0
SQLAlchemy~=2.0.15
starlette~=0.26.1
tenacity~=8.2.3
typing_extensions==4.6.2
uvicorn~=0.22.0
wrapt==1.15.0